+++ recipes/default.rb
@@ -17,9 +17,7 @@ platform_version = node['platform_version']
 raise("#{platform_family}/#{platform}/#{platform_version} is not supported by the default recipe") \
   unless platform_family?('rhel') &&
          node['yum-nginx']['rhel']['supported-versions']
-         .select { |_version, is_included| is_included }
-         .keys
-         .include?(platform_version.to_i.to_s)
+         .select { |_version, is_included| is_included }.key?(platform_version.to_i.to_s)
 
 node['yum-nginx']['repos'].each do |repo, value|
   next unless value['managed']