+++ test/fixtures/cookbooks/test/metadata.rb
@@ -1,8 +1,5 @@
 # frozen_string_literal: true
 
 name 'test'
-maintainer 'St. Isidore de Seville'
-maintainer_email 'st.isidore.de.seville@gmail.com'
-license 'MIT'
 description 'Installs/Configures test'
 version '0.1.0'