+++ CHANGELOG.md
@@ -9,8 +9,9 @@ cookbook.
 - Dang Nguyen (haidangwa@gmail.com)
   - set default make_cache property of yum repository to true
   - set explicit constraints in Gemfile
+  - updated rack, `~> 2.0.6` (CVE-2018-16470, CVE-2018-16471)
   - dropped centos 5 and centos 6 i386 support
-  - set chef_version to '> 12.14', '< 14.0' in metadata
+  - set chef_version to '> 12.14' in metadata
   - no longer requires yum cookbook dependency (use chef 12.14's native yum_repository resource)
   - install Chef 13 in Test Kitchen environment
   - Fixed Issue [#8](https://github.com/st-isidore-de-seville/cookbook-yum-nginx/issues/8)